requirements.txt
numpy
pillow
faiss-cpu
ollama